+++ client/src/features/results/ConceptSnippet.tsx
@@ -16,6 +16,7 @@ import StyledSnippetTitle from '../../styles/features/results/SnippetTitle'
 import RecordLink from '../common/RecordLink'
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
 import theme from '../../styles/theme'
+import { DEFAULT_PAGE_LENGTH } from '../../config/searchTypes'
 
 import SnippetHeader from './SnippetHeader'
 
@@ -33,7 +34,7 @@ const ConceptSnippet: React.FC<IProps> = ({
   view,
   totalResults,
   index,
-  pageLength = 20,
+  pageLength = DEFAULT_PAGE_LENGTH,
   titleOfTabbedContent,
 }) => {
   const [isMobile, setIsMobile] = useState<boolean>(