+++ client/src/features/results/ObjectSnippet.tsx
@@ -22,6 +22,7 @@ import {
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
 import theme from '../../styles/theme'
 import config from '../../config/config'
+import { DEFAULT_PAGE_LENGTH } from '../../config/searchTypes'
 
 import ProductionSnippet from './ProductionSnippet'
 import SnippetHeader from './SnippetHeader'
@@ -40,7 +41,7 @@ const ObjectSnippet: React.FC<ISearchData> = ({
   view,
   totalResults,
   index,
-  pageLength = 20,
+  pageLength = DEFAULT_PAGE_LENGTH,
   titleOfTabbedContent,
 }) => {
   const [isMobile, setIsMobile] = useState<boolean>(