+++ client/src/features/results/Paginator.tsx
@@ -1,10 +1,11 @@
 import { useMemo } from 'react'
 
 import { getPagination } from '../../lib/util/paginationHelper'
+import { DEFAULT_PAGE_LENGTH } from '../../config/searchTypes'
 
 export const Paginator = ({
   estimate = 0,
-  pageSize = 20,
+  pageSize = DEFAULT_PAGE_LENGTH,
   siblingCount = 2,
   currentPage = 1,
 }): Array<string | number> => {